+++ tests/conftest.py
@@ -44,7 +44,11 @@ def backend1(requests_mock, mbldr) -> str:
     domain = "https://b1.test/v1"
     # TODO: how to work with different API versions?
     requests_mock.get(
-        domain + "/", json=mbldr.capabilities(title="Dummy Federation One", description="Welcome to Federation One.")
+        domain + "/",
+        json=mbldr.capabilities(
+            title="Dummy Federation One",
+            description="Welcome to Federation One.",
+        ),
     )
     requests_mock.get(domain + "/credentials/oidc", json=mbldr.credentials_oidc())
     requests_mock.get(domain + "/processes", json=mbldr.processes(*_DEFAULT_PROCESSES))
@@ -54,7 +58,12 @@ def backend1(requests_mock, mbldr) -> str:
 @pytest.fixture
 def backend2(requests_mock, mbldr) -> str:
     domain = "https://b2.test/v1"
-    requests_mock.get(domain + "/", json=mbldr.capabilities(title="Dummy The Second"))
+    requests_mock.get(
+        domain + "/",
+        json=mbldr.capabilities(
+            title="Dummy The Second",
+        ),
+    )
     requests_mock.get(domain + "/credentials/oidc", json=mbldr.credentials_oidc())
     requests_mock.get(domain + "/processes", json=mbldr.processes(*_DEFAULT_PROCESSES))
     return domain